Create Python_ShippingCost_MiniProject
This is my original code for a Python mini project that involves calculating shipping costs from Codecademy Pro. Successfully writing the code for this project demonstrates mastery of writing basic Python functions, basic if/elif/else logic in Python, and returning different values in a function based on relational operators.