Create Python_InventoryManagement
This is my original code for the Codecademy Mini-Project managing inventory for a fictional flower shop. It uses the Pandas package and requires command of loading and managing CSVs including changing column names and performing column calculations, as well as adding new columns based on those calculations using lambda functions.